requirements.txt
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.5.2
astor==0.8.1
astroid==2.4.2
atomicwrites==1.4.1
attrs==19.3.0
autoflake==1.4
Babel==2.9.0
bandit==1.7.0
black==20.8b1
boto3==1.17.33
botocore==1.20.33
cachelib==0.1.1
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
click==7.1.2
colorama==0.4.5
coverage==5.5
cryptography==3.4.8
cs50==6.0.2
cycler==0.10.0
darglint==1.7.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.15
django-cors-headers==3.5.0
django-environ==0.4.5
django-heroku==0.3.1
django-rest-knox==4.1.0
django-simple-history==3.0.0
django-storages==1.11.1
django-taggit==1.5.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
doc8==0.8.1
docutils==0.16
eradicate==1.0
facebook-sdk==3.1.0
filelock==3.0.12
flake8==3.9.0
flake8-2020==1.6.0
flake8-bandit==2.1.2
flake8-broken-line==0.2.1
flake8-bugbear==19.8.0
flake8-builtins==1.5.3
flake8-commas==2.0.0
flake8-comprehensions==3.4.0
flake8-debugger==3.2.1
flake8-docstrings==1.6.0
flake8-eradicate==0.3.0
flake8-fixme==1.1.1
flake8-isort==3.0.1
flake8-logging-format==0.6.0
flake8-mutable==1.2.0
flake8-polyfill==1.0.2
flake8-quotes==2.1.2
flake8-rst-docstrings==0.0.12
flake8-string-format==0.2.3
flake8-variables-names==0.0.3
Flask==1.1.2
Flask-Session==0.3.2
gitdb==4.0.5
GitPython==3.1.14
gunicorn==20.0.4
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
isort==4.3.21
itsdangerous==1.1.0
jellyfish==0.8.2
Jinja2==2.11.2
jmespath==0.10.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.2.2
markdown2==2.3.9
MarkupSafe==1.1.1
martor==1.6.4
matplotlib==3.3.3
mccabe==0.6.1
mypy-extensions==0.4.3
mysql-connector==2.2.9
mysql-connector-python==8.0.20
numpy==1.23.2
packaging==20.9
pathspec==0.8.1
pbr==5.5.1
pep8-naming==0.9.1
pexpect==4.8.0
Pillow==8.0.1
pluggy==0.13.1
protobuf==3.12.2
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.6.0
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pydocstyle==5.1.1
pyflakes==2.3.0
Pygments==2.8.1
PyJWT==1.7.1
pylint==2.5.3
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
python-dateutil==2.8.1
python-dev-tools==2020.9.10
python-dotenv==0.17.0
pytz==2020.1
pyupgrade==2.11.0
PyYAML==5.3.1
regex==2021.3.17
requests==2.24.0
restructuredtext-lint==1.3.2
s3transfer==0.3.6
six==1.15.0
smmap==3.0.5
snowballstemmer==2.1.0
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.22
sqlparse==0.3.1
stevedore==3.3.0
termcolor==1.1.0
testfixtures==6.17.1
tokenize-rt==4.1.0
toml==0.10.1
tox==3.23.0
tox-travis==0.12
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzdata==2022.2
urllib3==1.25.9
virtualenv==20.2.0
wemake-python-styleguide==0.14.1
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
XlsxWriter==1.3.7
